Fix after cherry-pick
